requirements.txt
Django==5.2.4
djangorestframework==3.16.0
python-decouple==3.8
requests==2.32.4
python-whois==0.9.5
dnspython==2.7.0
email-validator==2.2.0
Pillow==10.4.0
django-cors-headers==4.6.0
celery==5.4.0
redis==5.2.0
gunicorn==23.0.0
psycopg2-binary==2.9.10